pull SizedAvatar out of User
